# doc/api.yaml
openapi: 3.0.3
servers:
- url: "http://localhost:3000"
info:
  title: WASAPhoto APIs
  description: |
    This OpenAPI document describes the apis for the WASAPhoto web application project.
  version: 3.0.3
...
paths:
...
  /users/{myUsername}/username:
    parameters: 
      - $ref: "#/components/parameters/myUsername"
    put: 
      security: 
        - bearerAuth: []
      tags: ["User"]
      summary: Modify your username
      description: |
        This function lets a user update their username. The new desired
        username must be sent in the requedt body.
      operationId: setMyUserName
      requestBody:
        content:
          application/json:
            schema: 
              $ref: "#/components/schemas/Username"
        required: true
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/User"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users/{myUsername}/bans/{bannedUsername}:
    parameters:
      - $ref: "#/components/parameters/myUsername"
      - $ref: "#/components/parameters/bannedUsername"
    delete:
      security: 
        - bearerAuth: []
      tags: ["Ban"]
      operationId: unbanUser
      summary: Unban a user
      description: |
        This function will remove a person from the ban list making it unbanned.
        The username of the person to unban must be specified in the path
      responses:
        '204':
          description: user unbanned successfully
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users/{myUsername}/follows/{followedUsername}:
    parameters: 
      - $ref: "#/components/parameters/myUsername"
      - $ref: "#/components/parameters/followedUsername"
    delete:
      security: 
        - bearerAuth: []
      tags: ["Follow"]
      operationId: unfollowUser
      summary: Stop following a user
      description: |
        This will be used to stop following a user by deleting it from the
        followed users list. The username of the person to unfollow must be
        specified in the path.
      responses:
        "204":
          description: User unfollowed successfully
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /users/{myUsername}/photos:
    parameters:
      - $ref: "#/components/parameters/myUsername"
    post:
      security: 
        - bearerAuth: []
      tags: ["Photo"]
      operationId: uploadPhoto
      summary: Upload a new photo
      description: |
        This function is used to upload a new photo. The image must be passed
        inside the request body.
      requestBody:
        content:
          image/*:
            schema:
              description: Image file
              type: object
              properties:
                 file:
                  description: the image file uploaded by the user
                  type: string
                  format: binary
                  minLength: 0
                  maxLength: 17179869184
      responses:
        '201':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotosList"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users/{myUsername}/photos/{photoID}/likes/{likeID}:
    parameters:
      - $ref: "#/components/parameters/myUsername"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/likeID"
    delete:
      security: 
        - bearerAuth: []
      tags: ["Like"]
      operationId: unlikePhoto
      summary: Unlike a photo
      description: | 
        Unlike a photo you previously liked by specifying the like id in the path
      responses:
        '200':
          description: Photo unliked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikesList"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
  /users/{myUsername}/photos/{photoID}/comments/{commentID}:
    parameters:
      - $ref: "#/components/parameters/myUsername"
      - $ref: "#/components/parameters/photoID"
      - $ref: "#/components/parameters/commentID"
    delete:
      security: 
        - bearerAuth: []
      tags: ["Comment"]
      operationId: uncommentPhoto
      summary: Delete your comment from a photo
      description: |
        Delete a comment you previously posted under a photo. The comment id and
        the photo id are specified in the path.
      responses:
        '200':
          description: comment delete successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsList"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          $ref: "#/components/responses/InternalServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: |
        The server cannot understand or process 
        correctly the request
    InternalServerError:
      description: |
        The server encountered an unexpected condition 
        that prevented it from fulfilling the request
    UnauthorizedError:
      description: |
        You don't have the permissio for this request
...
  schemas:
    Username:
      description: "The unique username string of an account"
      type: object
      properties:
        name:
          description: the username string
          type: string
          example: ObiWan_Kenobi
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
    Photo:
      description: |
        A photo object containing the photo id, the number of
        likes, 
      type: object
      properties:
        photoID:
          $ref: "#/components/schemas/ID"
        photoData:
          description: the actual image data
          type: string
          format: binary
          minLength: 0
          maxLength: 17179869184
        whoPosted:
          $ref: "#/components/schemas/Username"
        date:
          description: date of upload
          type: string
          format: date-time
          example: '2024-05-04T09:12:28Z'
          minLength: 1
          maxLength: 9999
        likesCount:
          $ref: "#/components/schemas/Count"
        commentsCount:
          $ref: "#/components/schemas/Count"
    Comment:
      description: | 
        A comment object with its id, its author and the text content
      type: object
      properties:
        commmentID:
          $ref: "#/components/schemas/ID"
        author:
          $ref: "#/components/schemas/Username"
        content:
          description: "the comment text content"
          type: string
          example: WOW! This is such a nice comment :)
          minLength: 0
          maxLength: 10000
          pattern: '^.+$'
    Like:
      description: | 
        A like object with an id, the username of the 
        person who liked and the id of the liked photo
      type: object
      properties:
        likeID:
          $ref: "#/components/schemas/ID"
        userWhoLiked:
          $ref: "#/components/schemas/Username"
        photoID:
           $ref: "#/components/schemas/ID"
    User:
      type: object
      title: User
      description: | 
        The user profile made of the username, the followers and 
        following counts, the number of posts
      properties:
        ID:
          allOf: 
            - $ref: "#/components/schemas/ID"
          readOnly: true
        Username:
          $ref: "#/components/schemas/Username"
        PostsCount:
          $ref: "#/components/schemas/Count"
...
    ID:
      description: |
        The integer id schema used for different objects like users, photos, comments,
        likes, bans and follows.
      type: integer
      example: 1234
      minimum: 0
      maximum: 1000000
    Count:
      description: "count schema used to count followers, comments, likes and posts"
      type: integer
      example: 42
      minimum: 0
      maximum: 1000000
    PhotosList:
      description: "The list of photos posted by a user"
      type: array
      items:
        $ref: "#/components/schemas/Photo"
      minItems: 0
      maxItems: 10000
...
    LikesList:
      description: "The list of likes on a photo"
      type: array
      items:
        $ref: "#/components/schemas/Like"
      minItems: 0
      maxItems: 10000
    CommentsList:
      description: "The list of comments under a photo"
      type: array
      items:
        $ref: "#/components/schemas/Comment"
      minItems: 0
      maxItems: 10000
  parameters:
    myUsername:
      schema:
        $ref: "#/components/schemas/Username"
      name: myUsername
      in: path
      required: true
      description: "the username of the logged in user"
...
    followedUsername:
      schema:
        $ref: "#/components/schemas/Username"
      name: followedUsername
      in: path
      required: true
      description: "The username of a user to unfollow."
    bannedUsername:
      schema:
        $ref: "#/components/schemas/Username"
      name: bannedUsername
      in: path
      required: true
      description: "banID: id of a user-bannedUsername object."
    photoID:
      schema:
        $ref: "#/components/schemas/ID"
      name: photoID 
      in: path
      required: true
      description: "photoID: id of a specific photo."
    commentID:
      schema:
        $ref: "#/components/schemas/ID"
      name: commentID
      in: path
      required: true
      description: "commentID: id of a comment under a photo."
    likeID:
      schema:
        $ref: "#/components/schemas/ID"
      name: likeID
      in: path
      required: true
      description: "likeID: id of a like on a photo."